Scanning Refactor V3 (#961)
* Fix _YARGSOURCE not using the explicit SortString constructor

* Fix YARGImage `!= null` comparisons

* Update VenueLoader to use string over DirectoryInfo

* Fix YARGImage source icon load

* Only hold the sort string form as the songsource key

* Fix ReplayContainer FixedArrayStream usage

* Configure .sng & con file "open file location" to pre-select the file

Currently windows and osx only.

* Fix song entry metadata accessor calls

* SongSearch: use correct namespaces for transforms

* Account for new `duplicate song` handling in music library menu

* SongContainer touchups

* Fix SortString JSONConverter

* Correctly update `All Songs` count on `Allow Duplicate Songs` settings change

* Fix error is `SongContainer.HasInstrument`

We can simplify that now.

+++ Assets/Script/Song/SongExport.cs
@@ -67,7 +67,7 @@ namespace YARG.Song
                     genre = RichTextUtils.StripRichTextTags(song.Genre),
                     charter = RichTextUtils.StripRichTextTags(song.Charter),
                     year = RichTextUtils.StripRichTextTags(song.UnmodifiedYear),
-                    songLength = song.SongLengthMilliseconds
+                    songLength = (ulong)song.SongLengthMilliseconds
                 });
             }
